# Optimizations for ARM Cortex-M4/7/33