This was more difficult than for ESP32, as ARM has hardware instructions for full div and sqrt, and no hardware estimates. But it was a fun challenge.

I developed a new approximation for fast_invsqrt() that is 25% faster and has 2x less error than the original (Quake) algorithm. It is similar, but uses exhaustive-search optimized coefs, and uses the extra degree of freedom to eliminate one of the multiplies (eg. coef constrained to 1.0).

I also replaced fast_inv() with a far more accurate one (~20 bits) that is sufficient to avoid artifacts, and still faster than a hardware division. How much depends on which CPU and compiler flags are used.

Here are error plots of the new approximations. Note that since the result exponent is exact (-E for 1/x, -E/2 for 1/sqrt(x)) the minimax error will just repeat over the full range.
![fast_inv](https://github.com/user-attachments/assets/7faa3c24-3fd6-441d-96a8-ae20f26aaea4)
![fast_rsqrtf](https://github.com/user-attachments/assets/e80cf642-7c51-4a86-a16b-dfd48f2a3a2d)