@vindar 2025-05-03
Hi Joe, 

Thanks again for the PR !
It is very impressive that you managed to improve still a bit on the quake trick :-) 

I just did some dirty testing to see how your change impact on a typical usage of the library, i.e. drawing a mesh... I realize it is by no means a good benchmark as too many factors influence the result (in particular instruction/data caching, optimization settings....) but I expect it shows what we may expect when using the library in 'real life'.

Here are the results for different boards (this is the rendering time, so less is best). I do not have an M0 currently at hand for testing this core variant. 

[tgx_benchmark.zip](https://github.com/user-attachments/files/20022912/tgx_benchmark.zip)

```
TEENSY 4.0 (M7), 600Mhz, -O"fastest with LTO"

before PR:
- Gouraud: (far) 2292us 	 (close) 2440us 	 (clipping) 1277us
- Texture: (far) 3195us 	 (close) 4526us 	 (clipping) 1485us
- flat   : (far) 1756us 	 (close) 1590us 	 (clipping) 1024us
- wire   : (far) 1692us 	 (close) 1764us 	 (clipping) 1026us
FINAL SCORE: 2005

after PR:
- Gouraud: (far) 2302us 	 (close) 2433us 	 (clipping) 1261us
- Texture: (far) 3232us 	 (close) 4365us 	 (clipping) 1466us
- flat   : (far) 1734us 	 (close) 1565us 	 (clipping) 984us
- wire   : (far) 1688us 	 (close) 1757us 	 (clipping) 1015us
FINAL SCORE: 1983


--------------------------------------------------------------------
TEENSY 3.6 (M4F), 180Mhz, -O"fastest with LTO"

before PR:
- Gouraud: (far) 12102us 	 (close) 11936us 	 (clipping) 5817us
- Texture: (far) 12465us 	 (close) 16949us 	 (clipping) 5853us
- flat   : (far) 9532us 	 (close) 8954us 	 (clipping) 5187us
- wire   : (far) 7094us 	 (close) 7417us 	 (clipping) 4471us
FINAL SCORE: 8981

after PR:
- Gouraud: (far) 11630us 	 (close) 11966us 	 (clipping) 6017us
- Texture: (far) 12546us 	 (close) 16727us 	 (clipping) 6061us
- flat   : (far) 9903us 	 (close) 9374us 	 (clipping) 5473us
- wire   : (far) 7136us 	 (close) 7439us 	 (clipping) 4482us
FINAL SCORE: 9062


--------------------------------------------------------------------
TEENSY 3.5 (M4F), 120Mhz, -O"fastest with LTO"

before PR:
- Gouraud: (far) 19554us 	 (close) 19775us 	 (clipping) 10216us
- Texture: (far) 20701us 	 (close) 28617us 	 (clipping) 10341us
- flat   : (far) 17544us 	 (close) 16199us 	 (clipping) 9627us
- wire   : (far) 14551us 	 (close) 15032us 	 (clipping) 8875us
FINAL SCORE: 15919

after PR:
- Gouraud: (far) 20124us 	 (close) 20254us 	 (clipping) 10627us
- Texture: (far) 21020us 	 (close) 28218us 	 (clipping) 10592us
- flat   : (far) 17991us 	 (close) 16666us 	 (clipping) 9924us
- wire   : (far) 14876us 	 (close) 15481us 	 (clipping) 9163us
FINAL SCORE: 16244


--------------------------------------------------------------------
TEENSY 3.2, (M4) 72Mhz, -O"fastest with LTO"

before PR:
- Gouraud: (far) 228424us 	 (close) 223633us 	 (clipping) 158656us
- Texture: (far) 256141us 	 (close) 345179us 	 (clipping) 160036us
- flat   : (far) 204527us 	 (close) 199652us 	 (clipping) 148142us
- wire   : (far) 142037us 	 (close) 133871us 	 (clipping) 114994us
FINAL SCORE: 192941

after PR:
- Gouraud: (far) 251543us 	 (close) 240997us 	 (clipping) 169781us
- Texture: (far) 283954us 	 (close) 445086us 	 (clipping) 172216us
- flat   : (far) 230432us 	 (close) 218992us 	 (clipping) 160392us
- wire   : (far) 154962us 	 (close) 144537us 	 (clipping) 123483us
FINAL SCORE: 216364


--------------------------------------------------------------------
PICO W (RP2040), 200Mhz, -Ofast

before PR:
- Gouraud: (far) 133691us 	 (close) 127257us 	 (clipping) 97467us
- Texture: (far) 150258us 	 (close) 181014us 	 (clipping) 97524us
- flat   : (far) 113375us 	 (close) 107247us 	 (clipping) 83668us
- wire   : (far) 81739us 	 (close) 79069us 	 (clipping) 67129us
FINAL SCORE: 109953

after PR:
- Gouraud: (far) 140422us 	 (close) 127015us 	 (clipping) 87674us
- Texture: (far) 143752us 	 (close) 174015us 	 (clipping) 89631us
- flat   : (far) 117583us 	 (close) 108083us 	 (clipping) 79625us
- wire   : (far) 80719us 	 (close) 78064us 	 (clipping) 67070us
FINAL SCORE: 107804


--------------------------------------------------------------------
PICO 2 (RP2350), 150Mhz, -Ofast (arch ARM)

before PR:
- Gouraud: (far) 17413us 	 (close) 16676us 	 (clipping) 10384us
- Texture: (far) 21165us 	 (close) 25373us 	 (clipping) 11405us
- flat   : (far) 14728us 	 (close) 12795us 	 (clipping) 9096us
- wire   : (far) 14182us 	 (close) 13880us 	 (clipping) 9553us
FINAL SCORE: 14720

after PR:
- Gouraud: (far) 16927us 	 (close) 16657us 	 (clipping) 10451us
- Texture: (far) 21470us 	 (close) 25471us 	 (clipping) 11535us
- flat   : (far) 13187us 	 (close) 12130us 	 (clipping) 8966us
- wire   : (far) 10627us 	 (close) 11185us 	 (clipping) 8115us
FINAL SCORE: 13893


--------------------------------------------------------------------
// ARDUINO 33 BLE (M4F) arduino default setting

before PR:
- Gouraud: (far) 50592us 	 (close) 54597us 	 (clipping) 29640us
- Texture: (far) 53924us 	 (close) 65641us 	 (clipping) 29944us
- flat   : (far) 47248us 	 (close) 43007us 	 (clipping) 28302us
- wire   : (far) 43376us 	 (close) 46441us 	 (clipping) 27095us
FINAL SCORE: 43317

after PR:
- Gouraud: (far) 51162us 	 (close) 54854us 	 (clipping) 29669us
- Texture: (far) 53840us 	 (close) 65427us 	 (clipping) 29626us
- flat   : (far) 47458us 	 (close) 43278us 	 (clipping) 28349us
- wire   : (far) 43653us 	 (close) 46678us 	 (clipping) 27063us
FINAL SCORE: 43421


--------------------------------------------------------------------
// ESP32 S2, (Xtensa, no FPU) 240MHz/80MHz, default settings

before PR:
- Gouraud: (far) 121855us 	 (close) 98768us 	 (clipping) 65804us
- Texture: (far) 155429us 	 (close) 151282us 	 (clipping) 70525us
- flat   : (far) 87726us 	 (close) 73683us 	 (clipping) 51774us
- wire   : (far) 52845us 	 (close) 50643us 	 (clipping) 42181us
FINAL SCORE: 85209

after PR:
- Gouraud: (far) 131089us 	 (close) 105592us 	 (clipping) 73454us
- Texture: (far) 165057us 	 (close) 182017us 	 (clipping) 78248us
- flat   : (far) 109180us 	 (close) 92907us 	 (clipping) 70837us
- wire   : (far) 56656us 	 (close) 53264us 	 (clipping) 44060us
FINAL SCORE: 96863
```

 As you can see, improvement varies depending on the board (the less powerful ones seem to benefit the last, perhaps because of pipeline congestion ?)

Even though, your improvements do not target ESP32, I also included the benchmark for ESP32 S2 since your previous assembly code does not apply for it. But unfortunately, it does not help :(

I also realized that my ``#define TGX_INLINE`` forcing inlining are not always beneficial: for some boards (but not all), it is better to let the compiler decide by himself what to inline...

So, I think I will merge your PR (your code is very nice and I love your tricks!) and subsequently I will clean up `misc.h`
to separate the preprocessor conditionals for each board. This will provide more granularity to selectively enable/disable optimizations (and also separate `TGX_INLINE` definition from the math speed optimization). 

Cheers !
Arvind