@vindar 2025-05-03
I just found another M7 board lying around...

```
WeAct STM32H743VIT6 (M7)  -O3 with LTO

before PR:
- Gouraud: (far) 3227us 	 (close) 3314us 	 (clipping) 1804us 
- Texture: (far) 3449us 	 (close) 4881us 	 (clipping) 1865us 
- flat   : (far) 2935us 	 (close) 2634us 	 (clipping) 1703us 
- wire   : (far) 2255us 	 (close) 2301us 	 (clipping) 1440us 
FINAL SCORE: 2650

after PR:
- Gouraud: (far) 3357us 	 (close) 3368us 	 (clipping) 1762us 
- Texture: (far) 3512us 	 (close) 4847us 	 (clipping) 1812us 
- flat   : (far) 3067us 	 (close) 2632us 	 (clipping) 1607us 
- wire   : (far) 2172us 	 (close) 2201us 	 (clipping) 1349us 
FINAL SCORE: 2640
```

Consistent with Teensy 4.0: not much change for M7.